+++ src/main/kotlin/blackjack/domain/card/Card.kt
@@ -1,0 +1,6 @@
+package blackjack.domain.card
+
+data class Card(
+    val suit: Suit,
+    val character: Character
+)
