+++ src/main/kotlin/blackjack/domain/card/Character.kt
@@ -1,0 +1,17 @@
+package blackjack.domain.card
+
+enum class Character(val value: Int) {
+    Ace(1),
+    Two(2),
+    Three(3),
+    Four(4),
+    Five(5),
+    Six(6),
+    Seven(7),
+    Eight(8),
+    Nine(9),
+    Ten(10),
+    Jack(10),
+    Queen(10),
+    King(10),
+}
