+++ src/main/kotlin/blackjack/domain/card/Suit.kt
@@ -1,0 +1,5 @@
+package blackjack.domain.card
+
+enum class Suit {
+    Spade, Diamond, Heart, Clover
+}
