+++ src/main/kotlin/blackjack/domain/player/PlayerState.kt
@@ -1,0 +1,5 @@
+package blackjack.domain.player
+
+enum class PlayerState {
+    Idle, Hit, Stay, Bust, Blackjack
+}
